+++ source/extensions/isaacsim.app.setup/premake5.lua
@@ -8,5 +8,5 @@ repo_build.prebuild_link {
     { "docs", ext.target_dir.."/docs" },
     { "data", ext.target_dir.."/data" },
     { "isaacsim", ext.target_dir.."/isaacsim" },
-    -- { "layouts", ext.target_dir.."/layouts" },
+    { "layouts", ext.target_dir.."/layouts" },
 }