+++ source/extensions/isaacsim.ros2.bridge/python/scripts/og_shortcuts/og_shortcut_menu.py
@@ -10,8 +10,10 @@
 
 import weakref
 
+import carb
 import omni.ext
 import omni.usd
+from isaacsim.core.utils.viewports import set_camera_view
 from isaacsim.gui.components.menu import make_menu_item_description
 from isaacsim.ros2.bridge.scripts.og_shortcuts.og_rtx_sensors import Ros2CameraGraph, Ros2RtxLidarGraph
 from isaacsim.ros2.bridge.scripts.og_shortcuts.og_utils import (
@@ -21,7 +23,16 @@ from isaacsim.ros2.bridge.scripts.og_shortcuts.og_utils import (
     Ros2OdometryGraph,
     Ros2TfPubGraph,
 )
-from omni.kit.menu.utils import MenuItemDescription, add_menu_items, remove_menu_items
+from isaacsim.storage.native.nucleus import get_assets_root_path
+from omni.kit.menu.utils import (
+    MenuItemDescription,
+    MenuLayout,
+    add_layout,
+    add_menu_items,
+    get_menu_layout,
+    remove_layout,
+    remove_menu_items,
+)
 
 
 class Extension(omni.ext.IExt):
@@ -29,39 +40,112 @@ class Extension(omni.ext.IExt):
         self.window_handle = None
         ros_og_menu = [
             make_menu_item_description(
-                ext_id, "ROS2 Camera", onclick_fun=lambda a=weakref.proxy(self): a._open_camera_sensor()
+                ext_id, "Camera", onclick_fun=lambda a=weakref.proxy(self): a._open_camera_sensor()
             ),
             make_menu_item_description(
-                ext_id, "ROS2 RTX Lidar", onclick_fun=lambda a=weakref.proxy(self): a._open_rtx_lidar_sensor()
+                ext_id, "RTX Lidar", onclick_fun=lambda a=weakref.proxy(self): a._open_rtx_lidar_sensor()
             ),
             make_menu_item_description(
-                ext_id, "ROS2 TF Publisher", onclick_fun=lambda a=weakref.proxy(self): a._open_pub_tf()
+                ext_id, "TF Publisher", onclick_fun=lambda a=weakref.proxy(self): a._open_pub_tf()
             ),
             make_menu_item_description(
                 ext_id,
-                "ROS2 Odometry Publisher",
+                "Odometry Publisher",
                 onclick_fun=lambda a=weakref.proxy(self): a._open_odometry_publisher(),
             ),
             make_menu_item_description(
-                ext_id, "ROS2 JointStates", onclick_fun=lambda a=weakref.proxy(self): a._open_joint_states_pubsub()
+                ext_id, "JointStates", onclick_fun=lambda a=weakref.proxy(self): a._open_joint_states_pubsub()
             ),
-            make_menu_item_description(ext_id, "ROS2 Clock", onclick_fun=lambda a=weakref.proxy(self): a._open_clock()),
+            make_menu_item_description(ext_id, "Clock", onclick_fun=lambda a=weakref.proxy(self): a._open_clock()),
             make_menu_item_description(
-                ext_id, "ROS2 Generic Publisher", onclick_fun=lambda a=weakref.proxy(self): a._open_rtf()
+                ext_id, "Generic Publisher", onclick_fun=lambda a=weakref.proxy(self): a._open_rtf()
             ),
         ]
 
         self._menu_items = [
             MenuItemDescription(
-                name="Common Omnigraphs",
+                name="ROS 2 Omnigraphs",
                 sub_menu=ros_og_menu,
             )
         ]
 
-        add_menu_items(self._menu_items, "Isaac Utils")
+        add_menu_items(self._menu_items, "Tools/Robotics")
+
+        ros_assets_sub_menu = [
+            MenuItemDescription(
+                name="Asset Browser", onclick_action=("isaacsim.asset.browser", "open_isaac_sim_asset_browser")
+            ),
+            MenuItemDescription(
+                name="Nova Carter",
+                onclick_fn=lambda a=weakref.proxy(self): a.create_asset(
+                    "/Isaac/Samples/ROS2/Robots/Nova_Carter_ROS.usd", "/nova_carter_ROS"
+                ),
+            ),
+            MenuItemDescription(
+                name="Leatherback",
+                onclick_fn=lambda a=weakref.proxy(self): a.create_asset(
+                    "/Isaac/Samples/ROS2/Robots/leatherback_ROS.usd", "/leatherback_ROS"
+                ),
+            ),
+            MenuItemDescription(
+                name="iw.hub ROS",
+                onclick_fn=lambda a=weakref.proxy(self): a.create_asset(
+                    "/Isaac/Samples/ROS2/Robots/iw_hub_ROS.usd", "/iw_hub_ROS"
+                ),
+            ),
+        ]
+
+        self._ros_assets_menu = [
+            MenuItemDescription(name="ROS 2 Assets", glyph="plug.svg", sub_menu=ros_assets_sub_menu)
+        ]
+
+        add_menu_items(self._ros_assets_menu, "Create")
+
+        # self.__ros_menu_layout = [
+        #     MenuLayout.Menu(
+        #         "Create",
+        #         [
+        #             MenuLayout.SubMenu(
+        #                 "ROS 2 Assets",
+        #                 [
+        #                     MenuLayout.Item("Asset Browser", source="Create/ROS 2 Assets/Asset Browser"),
+        #                     MenuLayout.Seperator("Examples"),
+        #                     MenuLayout.Item("Room", source="Create/ROS 2 Assets/Room"),
+        #                     MenuLayout.Item("Room 2", source="Create/ROS 2 Assets/Room 2"),
+        #                 ],
+        #             ),
+        #         ]
+        #     )
+        # ]
+        # add_layout(self.__ros_menu_layout)
+
+    def create_asset(self, usd_path, stage_path, camera_position=None, camera_target=None):
+
+        self._assets_root_path = get_assets_root_path()
+        if self._assets_root_path is None:
+            carb.log_error("Could not find Isaac Sim assets folder")
+            return
+
+        path_to = omni.kit.commands.execute(
+            "CreateReferenceCommand",
+            usd_context=omni.usd.get_context(),
+            path_to=stage_path,
+            asset_path=self._assets_root_path + usd_path,
+            instanceable=False,
+        )
+
+        carb.log_info(f"Added reference to {stage_path} at {path_to}")
+
+        if camera_position is not None and camera_target is not None:
+            set_camera_view(camera_position, camera_target)
+
+        pass
 
     def on_shutdown(self):
-        remove_menu_items(self._menu_items, "Isaac Utils")
+        remove_menu_items(self._menu_items, "Tools/Robotics")
+        remove_menu_items(self._ros_assets_menu, "Create")
+        # remove_layout(self.__ros_menu_layout)
+
         if self.window_handle:
             self.window_handle.visible = False