+++ source/extensions/isaacsim.asset.browser/docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [1.1.2] - 2024-11-01
+### Added
+- Registered action of opening asset browser
+
 ## [1.1.1] - 2024-11-01
 ### Changed
 - Rename browser