+++ source/extensions/isaacsim.robot.manipulators.ui/docs/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [2.1.0] - 2024-11-01
+### Changed
+- Menu name/location
+
 ## [2.0.1] - 2024-10-24
 ### Changed
 - Updated dependencies and imports after renaming